Move new app initialization out of main
